Custom Data Generators
Rudimentary implementation of custom data generators, tags have been moved to purely JSON, mutations read JSON data and adjust on reload, moving data from mutation classes to JSON files for data pack usage

+++ src/main/java/com/lenin/warpstonemod/common/mutations/attributes/IAttributeSource.java
@@ -5,7 +5,7 @@ import net.minecraft.entity.ai.attributes.AttributeModifier;
 import java.util.UUID;
 
 public interface IAttributeSource {
-	double getAttributeValue();
+	float getAttributeValue();
 	String getAttributeName();
 
 	void applyModifier (AttributeModifier source);
